+++ pkg/artifactory/provider/framework.go
@@ -16,6 +16,9 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	datasource_artifact "github.com/jfrog/terraform-provider-artifactory/v12/pkg/artifactory/datasource/artifact"
 	datasource_repository "github.com/jfrog/terraform-provider-artifactory/v12/pkg/artifactory/datasource/repository"
+	datasource_local "github.com/jfrog/terraform-provider-artifactory/v12/pkg/artifactory/datasource/repository/local"
+	datasource_remote "github.com/jfrog/terraform-provider-artifactory/v12/pkg/artifactory/datasource/repository/remote"
+	datasource_virtual "github.com/jfrog/terraform-provider-artifactory/v12/pkg/artifactory/datasource/repository/virtual"
 	"github.com/jfrog/terraform-provider-artifactory/v12/pkg/artifactory/resource/artifact"
 	"github.com/jfrog/terraform-provider-artifactory/v12/pkg/artifactory/resource/configuration"
 	"github.com/jfrog/terraform-provider-artifactory/v12/pkg/artifactory/resource/lifecycle"
@@ -343,6 +346,9 @@ func (p *ArtifactoryProvider) DataSources(_ context.Context) []func() datasource
 	return []func() datasource.DataSource{
 		datasource_repository.NewRepositoriesDataSource,
 		datasource_artifact.NewFileListDataSource,
+		datasource_local.NewLocalHexRepositoryDataSource,
+		datasource_remote.NewRemoteHexRepositoryDataSource,
+		datasource_virtual.NewVirtualHexRepositoryDataSource,
 	}
 }